Adding StudentService as well as repositories both for students and links between students and course

+++ src/CourseApi.Models/Student.cs
@@ -7,7 +7,7 @@ namespace CourseApi.Models
 {
     public class Student
     {
-        public int SSN { get; set; }
+        public long SSN { get; set; }
         public string Name { get; set; }
     }
 }
